FriendSearchService/friend/serializers.py
from rest_framework import fields, serializers

from friend.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('id', 'username', 'email')
