pyproject.toml
[tool.poetry]
name = "friendsearchservice"
version = "0.1.0"
description = ""
authors = ["allozo <allozo@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
